Feature/room options

+++ app/src/main/java/com/mo_chatting/chatapp/presentation/home/HomeFragment.kt
@@ -1,6 +1,7 @@
 package com.mo_chatting.chatapp.presentation.home
 
 import android.app.Activity
+import android.app.AlertDialog
 import android.content.ContentValues
 import android.content.Intent
 import android.os.Bundle
@@ -59,6 +60,7 @@ class HomeFragment : MyFragmentParent(), MyDialogListener, MyRenameDialogListene
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        while (viewModel.firebaseAuth.currentUser==null){}
         CoroutineScope(Dispatchers.IO).launch {
             setUserViews()
         }
@@ -94,7 +96,7 @@ class HomeFragment : MyFragmentParent(), MyDialogListener, MyRenameDialogListene
         withContext(Dispatchers.Main) {
             Glide.with(requireContext())
                 .load(uri)
-                .error(R.drawable.mohamed)
+                .error(R.drawable.ic_profile)
                 .override(500, 400)
                 .into(binding.profile)
         }
@@ -110,27 +112,54 @@ class HomeFragment : MyFragmentParent(), MyDialogListener, MyRenameDialogListene
         }
     }
 
+
     private fun setupRecyclerView() {
-        adapter = HomeRoomAdapter(viewModel.roomsList.value!!,
-            HomeRoomAdapter.OnRoomClickListener { room, position ->
-                onRoomClick(room, position)
-            },
-            HomeRoomAdapter.OnLongClickListener { room, position ->
-                onRoomLongClick(room, position)
-                false
-            },
-            HomeRoomAdapter.OnRoomDeleteClickListener { room, position ->
-                deleteRoom(room)
-            }
-        )
-        binding.rvHome.adapter = adapter
-        binding.rvHome.layoutManager = LinearLayoutManager(requireActivity())
+
+            adapter = HomeRoomAdapter(
+                viewModel.roomsList.value!!,
+                viewModel.firebaseAuth.currentUser!!.uid,
+                HomeRoomAdapter.OnRoomClickListener({ room, position ->
+                    onRoomClick(room, position)
+                }, { room, position ->
+                    onRoomLongClick(room, position)
+                    false
+                }, { room, position ->
+                    deleteRoom(room)
+                }, { room, position ->
+                    editRoom(room, position)
+                }, { room, position ->
+                    pinRoom(room, position)
+                }
+                )
+            )
+            binding.rvHome.adapter = adapter
+            binding.rvHome.layoutManager = LinearLayoutManager(requireActivity())
+
+    }
+
+    private fun editRoom(room: Room, position: Int) {
+        showEditRoomDialog(room, position)
+    }
+
+    private fun showEditRoomDialog(room: Room, position: Int) {
+        val editRoomDialog = CreateRoomDialog(this, true, room)
+        editRoomDialog.show(requireActivity().supportFragmentManager, null)
     }
 
     private fun deleteRoom(room: Room) {
-        CoroutineScope(Dispatchers.IO).launch {
-            viewModel.deleteRoom(room)
+        val builder = AlertDialog.Builder(context)
+        builder.setTitle("Do you want to leave?")
+        builder.setPositiveButton("Yes") { dialog, which ->
+            CoroutineScope(Dispatchers.IO).launch {
+                viewModel.deleteRoom(room)
+                dialog.dismiss()
+            }
         }
+        builder.setNegativeButton("No") { dialog, which ->
+            dialog.dismiss()
+        }
+        val dialog = builder.create()
+        dialog.show()
     }
 
     private fun onRoomLongClick(room: Room, position: Int) {
@@ -139,6 +168,10 @@ class HomeFragment : MyFragmentParent(), MyDialogListener, MyRenameDialogListene
 
     private fun onRoomClick(room: Room, position: Int) {
         findNavController().navigate(HomeFragmentDirections.actionHomeFragmentToChatFragment(room))
+    }
+
+    private fun pinRoom(room: Room, position: Int) {
+        showToast("not yet")
     }
 
     private fun setOnClicks() {
@@ -250,6 +283,12 @@ class HomeFragment : MyFragmentParent(), MyDialogListener, MyRenameDialogListene
     override fun onDataPassed(room: Room) {
         CoroutineScope(Dispatchers.IO).launch {
             viewModel.createNewRoom(room)
+        }
+    }
+
+    override fun onRoomEditPassed(room: Room) {
+        CoroutineScope(Dispatchers.IO).launch {
+            viewModel.updateRoom(room)
         }
     }
